@@ -30,10 +30,10 @@
 </template>
 
 <script>
-import eking from "@/lib/eking"
+
 import sweet from "sweetalert"
 import copy from "copy-to-clipboard"
-import $ from "jquery"
+// import $ from "jquery"
 
 export default{
 	name: 'formRadio',
