@@ -21,7 +21,7 @@
 import eking from "@/lib/eking"
 import sweet from "sweetalert"
 import copy from "copy-to-clipboard"
-import $ from "jquery"
+
 
 export default{
 	name: 'formCheckbox',
